Remove internal optic flavor modules if they're not needed
