+++ optics-core/src/Optics/Iso.hs
@@ -12,7 +12,7 @@ module Optics.Iso
   , uncurried
   , flipped
   , Swapped(..)
-  -- * Re-exprots
+  -- * Re-exports
   , module Optics.Optic
   )
   where
@@ -21,9 +21,35 @@ import Data.Tuple
 import Data.Bifunctor
 import Data.Coerce (Coercible, coerce)
 
-import Optics.Internal.Iso
+import Optics.Internal.Concrete
 import Optics.Internal.Optic
+import Optics.Internal.Profunctor
 import Optics.Optic
+
+-- | Type synonym for a type-modifying iso.
+type Iso s t a b = Optic An_Iso NoIx s t a b
+
+-- | Type synonym for a type-preserving iso.
+type Iso' s a = Optic' An_Iso NoIx s a
+
+-- | Explicitly cast an optic to an iso.
+toIso :: Is k An_Iso => Optic k is s t a b -> Optic An_Iso is s t a b
+toIso = castOptic
+{-# INLINE toIso #-}
+
+-- | Build an iso from a pair of inverse functions.
+iso :: (s -> a) -> (b -> t) -> Iso s t a b
+iso f g = Optic (dimap f g)
+{-# INLINE iso #-}
+
+-- | Extract the two components of an isomorphism.
+withIso :: Is k An_Iso => Optic k is s t a b -> ((s -> a) -> (b -> t) -> r) -> r
+withIso o k = case getOptic (toIso o) (Exchange id id) of
+  Exchange sa bt -> k sa bt
+{-# INLINE withIso #-}
+
+----------------------------------------
+-- Isomorphisms
 
 -- | This can be used to lift any 'Iso' into an arbitrary 'Functor'.
 mapping
