+++ optics-core/src/Optics/Prism.hs
@@ -17,9 +17,43 @@ module Optics.Prism
 
 import Data.Bifunctor
 
+import Optics.Internal.Concrete
 import Optics.Internal.Optic
-import Optics.Internal.Prism
+import Optics.Internal.Profunctor
 import Optics.Optic
+
+-- | Type synonym for a type-modifying prism.
+type Prism s t a b = Optic A_Prism NoIx s t a b
+
+-- | Type synonym for a type-preserving prism.
+type Prism' s a = Optic' A_Prism NoIx s a
+
+-- | Explicitly cast an optic to a prism.
+toPrism :: Is k A_Prism => Optic k is s t a b -> Optic A_Prism is s t a b
+toPrism = castOptic
+{-# INLINE toPrism #-}
+
+-- | Build a prism from a constructor and a matcher.
+prism :: (b -> t) -> (s -> Either t a) -> Prism s t a b
+prism construct match = Optic $ dimap match (either id construct) . right'
+{-# INLINE prism #-}
+
+-- | This is usually used to build a 'Prism'', when you have to use an operation
+-- like 'Data.Typeable.cast' which already returns a 'Maybe'.
+prism' :: (b -> s) -> (s -> Maybe a) -> Prism s s a b
+prism' bs sma = prism bs (\s -> maybe (Left s) Right (sma s))
+{-# INLINE prism' #-}
+
+withPrism
+  :: Is k A_Prism
+  => Optic k is s t a b
+  -> ((b -> t) -> (s -> Either t a) -> r)
+  -> r
+withPrism o k = case getOptic (toPrism o) (Market id Right) of
+  Market construct match -> k construct match
+{-# INLINE withPrism #-}
+
+----------------------------------------
 
 -- | Use a 'Prism' to work over part of a structure.
 aside :: Is k A_Prism => Optic k is s t a b -> Prism (e, s) (e, t) (e, a) (e, b)
