+++ optics-core/src/Optics/Review.hs
@@ -4,12 +4,31 @@ module Optics.Review
   , toReview
   , review
   , unto
-  , ReversibleOptic(..)
   , module Optics.Optic
   )
   where
 
+import Optics.Internal.Bi
 import Optics.Internal.Optic
-import Optics.Internal.Re
-import Optics.Internal.Review
+import Optics.Internal.Profunctor
+import Optics.Internal.Tagged
+import Optics.Internal.Utils
 import Optics.Optic
+
+-- | Type synonym for a review.
+type Review t b = Optic' A_Review NoIx t b
+
+-- | Explicitly cast an optic to a review.
+toReview :: Is k A_Review => Optic k is s t a b -> Optic A_Review is s t a b
+toReview = castOptic
+{-# INLINE toReview #-}
+
+-- | Apply a review.
+review :: Is k A_Review => Optic' k is t b -> b -> t
+review o = unTagged #. getOptic (toReview o) .# Tagged
+{-# INLINE review #-}
+
+-- | An analogue of 'to' for review.
+unto :: (b -> t) -> Review t b
+unto f = Optic (lphantom . rmap f)
+{-# INLINE unto #-}
